Former-commit-id: 8926fbc7a4d7ec095b255f75a41336c0385f65a4
Former-commit-id: e6b35ae6a8d1b9aa7875a7bd2b06e3361e7a65d6

+++ NINA/ViewModel/FramingAssistant/FramingAssistantVM.cs
@@ -552,9 +552,9 @@ namespace NINA.ViewModel.FramingAssistant {
             }
         }
 
-        private int _focalLength;
+        private double _focalLength;
 
-        public int FocalLength {
+        public double FocalLength {
             get {
                 return _focalLength;
             }
